Change LCL implementation to be more friendly.
This changes from using the minimize() function from scipy.optimize to a
method that uses fixed point iteration to find convergence. Not only is
this 30x faster for a single point, but this is both easier and
infinitely faster to use for arrays. Also correct some temperature unit
issues with examples and parcel_profile() so that we really are using
Kelvin for input as specified in the docs.