Modify circumcircle radius calculation
Found an alternative calculation that uses the area of a triangle which
allows for a more intuitive check for a valid triangle.  Namely, if area
of the triangle is zero, do not calculate radius.  We could also use
this to check for triangles with areas that are nearly zero in the
future.

Also removed redundant code from circumcenter.