ENH: Support reading Wyoming upper air archive. (Fixes #9)
This adds one function to hit the archive, and another to parse the
data. This helps separate out the remote-access functionality (which I
don't want to include in automated testing) and parsing code which can
be tested and used with our local test data.

We also put back the partial line in the local sounding data so that
they can be fully parsed by the new code.