MNT: Tighten up/remove noqa.
We can specify particular errors to suppress. This better documents the
need for the noqa, as well as keeps from suppressing any other problems
that may come along. There are also a few noqa cases that are no longer
needed. we now ignore _version, and apparently flake8 doesn't alert for
_ being unused.