MNT: Refactor GINI tests.
Instead of duplicating the procedure used to check out the file, pull
out the reference data and make a parameterized test. This also catches
another nose->pytest conversion glitch where the size of the output data
stopped being checked.

This adds a method to NamedStruct allowing it to just construct the
tuple from values. This is used to make it easy to test the results of
parsing.