BUG: Fix ordering of GINI y-coordinates
GINI raster data come ordered with the top (i.e. highest latitude) row
first. Before this fix, y coordinates were calculated increasing from
the lower left corner--thus having the y-coordinates flipped from the
ordering of rows in the image. This was not enountered frequently due to
the use of imshow() with extents and origin.

This reverses the y coordinates and adjusts tests to match.