ENH: Clean up array ordering in kinematics. (Fixes #389)
By fixing the array ordering of these functions as xy, users are most
frequently needing a large amount of transposes. Instead, we make it
possible to specify the ordering. For now, we keep the same default, but
issue a warning because we will change in 0.6.

Add a decorator to change all incoming arrays as necessary based on the
dim_order kwarg, and make all calculations assume y, x order so that
they operate most efficiently on the common case.