MNT: Refactor lcl to use scipy.optimize
Instead of our own fixed point iteration code, use the code in
scipy.optimize. While this is measurably slower, we're still only
talking tens of milliseconds for 90000-element arrays. This slowdown is
outweighed by leveraging a re-usable algorithm.

This also fixes a minor bug in lcl where we were returning the dewpoint
corresponding to the next-to-last pressure value from the iteration. In
practice, we're talking about differences in the thousandth of a degree.