BUG: Fix get_layer with float32 data
It turns out the numpy.interp() always returns float64, no matter the
input (except for complex). This causes problems when we try to append
the result of the interpolation to our float32 data, causing the array
to upcast. This in turn causes, depending on the truncation, checks to
consider the lowest layer to seem like it's outside the bounds of the
data.