fix to erroneous/misleading theta-e formula
The formula as it stood is quite approximate, and missed a crucial aspect: theta-e depends on moisture content!
The formula as it stands involves T,p at the LCL, not just T,p. I clarified that. 

But it would be better to add a more accurate theta-e formula to thermo.py, like the Bolton (1980) formula for numerical accuracy, it can be found at 
https://en.wikipedia.org/wiki/Equivalent_potential_temperature