Update wx_code_map in wx_symbols.pl
Modified the wx_code_map to include further entries.  There is a discrepancy found in the previous dictionary as well where 'UP' was set to code 16, but that is only when the station is automated.  Automated station symbols come from a different table than the one we are currently mapping to.  Read more at: https://github.com/Unidata/MetPy/files/1151142/485_Vol_I_en.pdf in pages II-4-3 and II-4-4.