Rename Convergence Functions to Divergence
This commit is a simple rename of the kinematic functions labeled
"convergence" with "divergence" instead, which is what is actually
being calculated. It updates the corresponding tests along with that. It
also corrects a few small comments where the string "convergence"
happened to show up when it should not have.