Replace "View source" with "Improve this page"
Modify the ReadTheDocs 'breadcrumbs.html' template to include this link.
The link in the docs will point to the Github source in "edit" mode (on
the master branch). If the page is autogenerated, it will point to a new
issue instead with the page name included in the title and some
instructions in the body.