ENH: Use linear search in isentropic_interpolation (Fixes #798)
Previously, this was using numpy.searchsorted(), which is only strictly
correct if the points are sorted/monotonic increasing. Theta can easily
violate this condition. Instead, switch to a linear search, which allows
also controlling which direction to search. This also seems to be faster
since we're no longer using apply_along_axis, but can rely on indexing
operations.