Let dimension coordinates take precedence
In working with the GOES16 data, the coordinate identification in
parse_cf fails because there are two coordinates with
"projection_x_coordinate" as their standard name. But, only one of them
is actually a dimension coordinate, and we would want that one to be
selected. This commit implements and tests for this.