Allow extra leading dimensions in gradient and kinematics calculations
Based on issues with some kinematics calculations generalizing to data
that has extra leading dimensions (such as time or vertical as commonly
encountered when working with xarray examples), this commit modifies the
default derivative axes in kinematics calculations to be based on
trailing axes (-2 and -1) rather than leading (0 and 1), to follow our
convention (and CF option) of (..., Y, X) dimension ordering. Numerous
tests are added to make sure that the affected calculations work with
extra dimensions (tested on 4-D data).

Additionally, the gradient function previously only supported the full
gradient of an array according to its description, but a useful case is
just obtaining the horizontal gradient along all levels or times. Handling
of the "axes" kwarg and was added to the _process_gradient_args function
to resolve this, so that "axes" becomes a kwarg of gradient and
laplacian. It must be less than or equal to in length compared to the
coordinates/deltas.