NARR Example CF Parsing
While I was working through examples, I noticed that the parse_cf
function, when applied to the full dataset of the NARR example, would
fail...the x and y coordinates would double in size, and many NaN's
appeared. This was because the lon and lat variables lacked a
grid_mapping attribute, and so, their x and y coordinates did not get
scaled from km to m like the other variables that did. Thus, they did
not line up when xarray tried to combine them back together in a
non-unit-aware fashion.

This commit fixes this by changing when _fixup_coords is called, so that
all x/y coordinates get scaled accordingly.