Make derivative calculations work with xr.DataArray
This commit modifies the current derivative calculations
(first_derivative, second_derivative, gradient, and laplacian) to work
with DataArrays themselves, rather than by conversion to (and output as)
pint Quantities. In order to support lat/lon non-uniform coordinates,
the lat_lon_grid_deltas function has moved from kinematics to tools to
be alongside the derivative calculations, and a helper function to make
it work nicely with a DataArray was added.

Some small fixes are also made to the xarray tutorial along the way.