Clean up exisiting workarounds
In order to work with data that had 'degrees_north' (or
similar), workarounds were needed in the past. With the unit update,
these are no longer needed.

Part of the remaining discussion involved how exactly to handle 'gpm'
non-standard units. A comment referencing the relevant issue has been
added to the tutorial section on the existing workaround, since the
issue needs further implementation discussion.