BUG: Don't use basicConfig for logging (Fixes #866)
What basicConfig does is install a root handler if none is already
installed. This breaks applications that also try to use it to set a
different level. It's completely unnecessary on Python 3, since by
default there is a handler that logs anything WARNING or higher to
stderr, which is really what we want. This isn't the case on Python 2,
so what we can do in that case is put a StreamHandler (which defaults to
stderr) for all of metpy (all submodules will defer to it by default).
WARNING is already the default logging level anyway.