Add unit- and axis-aware selection to the metpy accessors
This commit adds the .loc indexer and .sel method to both the DataArray
and Dataset metpy accessors to allow for unit-aware indexing/selection.
For DataArrays (for which axis types are uniquely determined), this also
allows selection via the axis name ('time', 'vertical', etc.) instead of
the coordinate name.

Also, adds this new unit-aware selection to the xarray tutorial.