Fix LFC calculation when positive area exists below LCL
Soundings can exist where there is positive area only below the
LCL. This commonly occurs when there is a superadiabatic lapse rate
near the ground. Other soundings may have positive area all the way
from the surface to an EL above the LCL. The current code conflated
these two situations, assuming that any positive area in the sounding
must mean the second case is at play. This commit makes sure that any
positive area that is found is above the LCL. If so, the LFC is set to
the LCL as before, but if not, there is no LFC (it is set
np.nan). Changes suggested by @dopplershift are incorporated.