Rework SkewT transform to work in pixel space (Fixes #209)
This changes to apply the skewing last, which fixes some issues caused
by working in normalized axes coordinates, like problems when the aspect
ratio of the plot wasn't square. This likely paves the way for better
handling of data aspect ratio for skewT now too. The key here is to use
matplotlib's transform framework to listen for changes to the axes bbox,
so that the transform can update to use the proper origin for the
skewing.