skewt subroutines updated to handle non-celcius
Updated the moist adiabats, dry adiabats, and the mixing line subroutines
to use the xaxis units. These routines originally used degrees celcius
even if the xaxis was not degrees celcius. This caused lines to be
plotted wrongly on any other units.

A similar issue may occur if units other than mbar are used..

Resolves: #1155