BUG: Fix Pooch getting data added in a PR
Pooch considers the environment variable as a base directory for data,
with versions inside. We want to use it as a way to point directly to
our copy of data. So we can just use this variable directly to set the
pooch path, just like we do looking relative to the location of
cbook.py.