Modify xarray coord id to use MetPy axes, including lat/lon
To resolve #1090, this commit changes the CF-axis-based labeling of
coordinate identification ('T', 'Z', 'Y', 'X'), to a more general
labeling system (currently with labels 'time', 'vertical', 'y',
'latitude', 'x', and 'longitude'). This also permits the needed changes
of #1090, which are adding lat/lon coord id and setting the stage for
x/y to be 1D only. Also, explicit tests are added for the following
cases in regards to coord id:

- 1D latitude/longitude (which should also be assigned as y/x)
- 2D latitude/longitude with existing 1D y/x coords
- 2D latitude/longitude without 1D y/x coords (deprecated assignment as
y/x as well)