Fix for skewt subroutines ignoring x axis units
Wet and dry adiabats, along with mixing lines, do not correctly
plot unless everything is in degrees celcius. This is especially
a problem when the plots show the wrong values without giving
any sort of warning. This fix makes the subroutines plot onto
units matching the xaxis or input temperature.
Checklist

Closes #1155
test added