MNT: Allow use of new coverage
Coverage 4.3.4 on Python 3.8 produces some strange results that are
breaking CodeCov, so might as well update. Also, we're currently using
newer coverage on Windows anyways because it did not like the syntax we
were using to request an old one anyways--and codecov does not appear to
have a problem with it.