BUG: Fix units appearing on station plot
So with pint's advances, it turns out `item()` no longer strips units.
This is #330 come back for revenge. The right answer is to just drop
units when calling `plot_parameter`. We don't want them, and there's no
format code that will drop units; while one could be added, this
wouldn't work with floats in general.