ENH: Add a custom GitHub Action for our CI
This adds a full suite of jobs to handle building docs on Linux, as well
as running tests on Linux (using PyPI) and on Windows/Mac using Conda.
Instead of relying on what's encoded in setup.py and a bunch of manual
additions (Travis) or an environment.yml (AppVeyor), we have requirement
and constraint files that get passed/swapped depending on what kind of
build we're running. This allows us to pin versions and hopefully let
dependabot manage the updates to these so that CI doesn't fail due to
the ecosystem updating.