MNT: Rename dependency files
Rename minimum.txt and prerelease.txt to Minimum and Prerelease so that
they are no longer picked up by Dependabot. We *never* want Dependabot
to update this file, and this is pretty much the only way to do it.
prerelease.txt would have been fine, but this matches Minimum and looks
better in the UI. Capitalize current.txt to match.