TST: Skip appropriate tests when no CartoPy present
This allows us to run the test suite without CartoPy. This adds two
global fixtures that allow tests that need either cartopy.crs or
cartopy.feature to have fixture paramters that provide it rather than
import the library--the tests are skipped if the import fails. There's
also a `needs_cartopy` decorator used to mark tests that need cartopy to
function (usually because of `parse_cf`).