BUG: Fix subsetting in PlotObs.obsdata
If we had a level, we were not applying any of the time windowing. This
changes to independently apply each of the subsetting operations we were
doing (time, level, station). We also now make use of time as the index
of the pandas DataFrame, and sort before subsetting. This changed the
output for some of our surface data.