MNT: Update tests due to change to PyProj
Necessary because we are now properly using the Earth ellipsoid that is
present in the data for lat/lon projection. Previously, we were using
CartoPy's PlateCarree, which doesn't support globes--thus when we got
the spheroid we used the CartoPy default, which is WGS84.

Was able to reproduce previous test values by overriding data ellipsoid
to specify WGS84. Updating the test values is better long-term than
putting a WGS84 hack into the data fixture.