ENH: Make CartoPy import more robust (Fixes #1550)
Instead of just importing cartopy.feature, explicitly import the two
names we need. The result is that on older Cartopy, where Scaler is
unavailable, we fail as if CartoPy is not installed rather than failing
when creating our map features--which allows users with old CartoPy to
still use e.g. SkewT.

No tests because there's not obvious way to do so.