BUG: Fix wet_bulb_temperature for saturated conditions (Fixes #1332)
Turns out that moist_lapse did not cope with being given a uniform array
of pressures (which while not useful, should just return an array full
of the original temperature). This caused the wet_bulb_temperature
calculation to fail when conditions were saturated. Simplify by using
the support for passing a reference level explicitly to moist_lapse, and
only request the value for our desired pressure.