BUG: Fix get_layer with more float32 data
We've fixed this before, but this case involves the paths taken when we
have no heights provided. Also the culprit here is that while float32 *
float -> float32 for arrays, for scalars this is float64. Apparently
(See numpy 17816). The easiest workaround for now is to force it out of
the scalar path.