ENH: Add GitHub Action for drafting a release
This action, upon closing a milestone, will create a new *draft*
release, choosing an appropriate tag name and title based on the title
of the milestone. It also can choose an appropriate branch for the tag
if that's indicated with 'branch: mybranch' in the milestone text. The
rest of the body of the release is auto-generated release notes from the
PRs and issues, as well as the list of contributors.