Added scalloped path effect
This commit adds a path effect to make lines drawn appear to be
scalloped. This was based off of modifying the matplotlib
TickedStroke path effect. This style of line is often used for
highlighting areas in SPC MCD products.

Places it in metpy.plots.patheffects module with patches for
versions of matplotlib < 3.2 in the metpy.plots._mpl module.