Remove misleading fallback to lat/lon CRS when only 2D lat/lon present
As reported on StackOverflow and in https://github.com/Unidata/MetPy/issues/1648, it is incorrect to fallback to a lat/lon CRS when a CF grid mapping is missing and 2D lat/lons are present (as this means we instead have data on a projected grid of unknown CRS and not data on a lat/lon grid). This PR changes the lat/lon fallback to only occur when the identified latitude coordinate is also considered a y coordinate and likewise longitude is considered x (so that we truly have a lat/lon grid represented).

This did require some additional tweaks to the grid_deltas_from_dataarray utility, which assumed that a metpy_crs was always present (which never was really the case, but is especially important to not assume now since we no longer [mistakenly] inject a lat/lon CRS in the 2D lat/lon case).