CI: Add environment to upload to PyPI
Set things up to use a GitHub environment for uploading to PyPI. This
treats the upload to PyPI as a deployment, and restricts the PyPI token
to this environment. Ideally, this would allow us to restrict the
environment to our tagged releases, but unfortunately the "branch
filter" rules only work on branches, not tags.

This was motivated by CodeCov having a security breach. Our PyPI token
doesn't seem to be at risk, but it was rotated anyway and this seems
like a nice extra layer to put in the way of it accidentally being
exposed/used.