BUG: Fix uses of unit multiplication internally
These are problematic with masked arrays. Also, using the Quantity()
constructor is faster across the board. Therefore, just get rid of all
the places where we multiply by units.

This also adds a test for get_layer where it was failing with masked
arrays, which is what originally motivated this.