MNT: Change some top-level imports to function-level (Fixes #1901)
This avoids importing metpy.calc and metpy.plots in the metar parsing
code, which improves import times for code that's only doing unrelated
I/O, as well as eliminates a warning when CartoPy is not installed.

The wx_code_map and wx_code_to_numeric function should eventually be
moved to metar.py, but that's an API change.