Don't fail on old or bad NCEI storm data
When reading NCEI storm data, don't fail on bad or old data, in
particular for timezones this is a concern.  Add test cases for all
timezones found in the NCEI storm data base, including typos and
erroneaus data.  Adapt the implementation of timezone interpretation to
use the 3-letter code (always available) rather than the 1-digit hour
offset (not always available).