BUG: Fix a couple issues with the code->font mapping
Auto wx symbols was accidentally starting in the high cloud symbols,
instead of skipping the first 4 codes. Also, pressure tendency
accidentally went up to 10 instead of 9 (0-8). Adjust the tables in the
docs to display 10 per row to make a nicer square image for the wx
symbols *and* match the WMO tables, which revealed the errors.