ENH: Improve weather symbol handling
This changes the current_weather mapping in two ways. First, alternate
symbols can be selected by using values >1000, where the thousands digit
denotes the alternate number. This allows us to have an array of codes
that display the proper single-precip-type symbol

Second, the first 50 symbols in the automated station symbol table are
mapped to 101-150. This allows us to handle METAR reports that encode
unknown precipitation type (UP) with the proper symbols.