Validate MapPanel attributes
Right now, typos in the names of attributes used by the declarative
plotting interface lead to strange errors that can be difficult to
debug.  In this commit, a new mixin class is defined to provide
validation of attribute names against a whitelist.

This approach was inspired by a Stack Overflow question:
__setattr__ versus __slots__ for constraining attribute creation...

and the nearest-string function was stolen from Rosetta Code.

Since this is a proof of concept, validation is only applied to the
MapPanel object.