CI: Limit amount of characters submitted from nightly logs
This fixes a problem where there was so much output in the test and doc
failure logs that it caused opening a new issue to fail with an API
error. The maximum characters for the body is 65536, so just cap each
log to 21k characters. If more info is necessary, the link to the
failing job is available.