Fixed failing tests and extended test coverage
I found the culprit - it's the keep_attrs kwarg.
For some reason the kwarg does not only keep the dataset's attributes (like insinuated in the documentation) but also the dataarrays' attributes.
This does not work for us, as we explicitly want to change the dataarrays' attributes (quantify removes 'units' attr and dequantify adds it).

I also extended the test coverage to include an actual unit in the test_ds_generic.
Otherwise (in the dimensionless case), tests/test_xarray.py:189 will never fail.