BUG: Fix resolution of interpolate-generated grids (Fixes #2319)
There was an off-by-one error in calculating the number of grid points
needed to meet the desired resolution. The previous code calculated the
number of steps, but did not include the starting point.

In order to avoid regenerating test data, the specified resolution in
the test was adjusted to match that actually given before.