Fix bug in GempakGrid
A non-existent attribute GDTM2 was referenced when creating the
Grid namedtuple. Since a majority of GEMPAK files do not define a
second time on grids, it was not detected. That is, GDT2 is rarely
defined and the logic statement gets terminated before actually checking
for GDTM2. The parameter name was changed to the correct GTM2.