BUG: Improve GEMPAK "thinned GFS" files (Fixes #2121)
The problem is that the original GEMPAK CED and MCD ([modified]
cylindrical equidistant) projections, while supported by PROJ, aren't
defined in the CF metadata spec (as of 1.10). This is a work-around that
makes __str__ not fail, and gives a better error, when CFProjection is
given a set of metadata it can't handle.