BUG: Fix interpolation with mixed types
When given a DataArray and a numpy array, neither with units, a Quantity
would be passed into the core calculation, triggering an exception on
`apply_along_axis`. Instead, handle trying to convert and drop units on
a "dimensionless" Quantity, since that's what we get for a DataArray
with no units.