Check for duplicate levels in isentropic_interpolation_as_dataset
Fixes #3309 by provided the user with a warning when interpolation to
a duplicate isentropic surface is requested.  This is likely a mistake
on the part of the user, but by issuing a warning rather than an
error, computations can proceed if the duplication was really
intended.

The code used here is an adaptation of the duplicate-level check found
in _parcel_profile_helper elsewhere in thermo.py.