BUG: Correct vertical coordinate assumption (Fixes #3315)
isentropic_interpolation_as_dataset was blindly assuming that the
vertical coordinate was pressure, resulting in a confusing unit error
when calling isentropic_interpolation. Check the dimensionality and
raise a clearer ValueError when this is not the case. Also allow users
to manually pass (as a keyword) an appropriate pressure array to use instead.