ENH: Make coordinate detection more robust (Fixes #3416)
This works around an issue where some ERA5 data had lat/lon with a
"positive" attribute that, according to the CF specification, should
indicate a vertical coordinate. We can fix this up by making sure each
coordinate variable is only associated with a single option, and making
use of Python's key-ordered dictionaries to prioritize the matching
order. Having lat be y and lon be x is already handled by a later block.