BUG: Fix turbulence kinetic energy calculation (Fixes #3756)
The turbulence kinetic energy calculation was incorrectly taking the
square root of the squares of the perturbation velocity components.  The
tests were also updated to use new, non-zero mean velocity components. A
new, separate test was added to testing the tke calculation when
perturbation velocities were supplied to the function.