BUG: Fix handling rotated pole (lat/lon) parsing
Instead of just assuming we should fall back to the geostationary
perspective_point_height support, check to see if that attribute is even
available, otherwise there's no point in trying.

This wasn't done in the first place because it wasn't clear there were
other projections with valid coordinates that couldn't be converted to
meters.